backend/requirements.txt
# general  modules
Django==4.2.5
psycopg2-binary==2.9.7

django-cors-headers==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow

# third party modules
python-decouple
django-filter
django-daisy
django-debug-toolbar

# test
pytest 
pytest-django 
factory_boy 
pytest-cov

# background task

celery
redis
django-celery-beat